Update gocam-overview.md
Adding a link to the GO-CAM site, as I couldn't find one.  Do we want this to open in a new tab? The rule is internal links don't, but this seems "outside" geneontology.org-- maybe because of the different header/footers, that may be a different conversation.  

so, {:target="blank"} ?