Update download-ontology.md
Rearranging so the Download links are at the top.  
Adding more words to make it clearer where each link goes.

Importantly, killing off mention of SVN- need to verify this is ok with @kltm but I think scrubbing the mention of SVN goes along with convos from about two years ago

Where was the suggested new home of Subsets?