Update literature.md
I'm assuming the stats for 
 experimental findings from over [160,000 published papers]
and 
 represented as over 875,000 experimentally-supported annotation

are from the stats? In which case I don't think (https://www.ncbi.nlm.nih.gov/pubmed/?term=loprovGeneOntol[SB]) is in the right place, maybe that number needs to be near `GO is referenced in tens of thousands of scientific papers.`


New section to add on the white paper link.  Is there another link of all go-written papers I can add on?