Update introduction-go-resource.md
Turns out we didn't really mention the Alliance yet, really, anywhere on the site (oops).  Also added a summary of GCBR- this is slanted toward focus on GBC instead of GCBR to get away from "GO as a resource".  Any help with phrasing is appreciated, or even if this should be on a different page (but if so, where?). @vanaukenk @pgaudet